Fixed metadata index bug in LocalIndex.ts
The UUID method wasn't called, causing an error when caching the metadata. 

I would love your thoughts on whether to keep this functionality or remove/clean it for increased simplicity. Granted no-one has used this in over a year because it wasn't working. I can see about doing that if you're willing.

Anyway, great work on this project!