remove Object.assign which duplicates added copy
Originally `Object.assign` was being used to create a copy of the object.
I replaced it in a previous commit with `structuredClone` since the copy
was not a deep one.

I keps the `Object.assign` call in case the other functionality it offers
was relevant, but seeing as `_update` is checked to be `undefined`, there
are no properties to keep in the call to assign.